+++ src/main/java/bsw/iron/barcode_server/entity/GroupConversion.java
@@ -18,7 +18,7 @@ public class GroupConversion {
 
    @Id
    @Column(name = "GROUP_ID")
-   private Integer idGroup;
+   private Long idGroup;
 
    @Column(name = "NAME_GROUP")
    private String nameGroup;
