+++ src/main/java/bsw/iron/barcode_server/repository/TestHeadRepository.java
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TestHeadRepository  extends JpaRepository<TestHead, Integer> {
+public interface TestHeadRepository extends JpaRepository<TestHead, Long> {
 
 // Здесь хранятся captionName и их idTestHead для TestValue
 
-/**
- * Получение всех наименований (caption) значений для передела 11690 в подчиненной таблице
- */
+    /**
+     * Получение всех наименований (caption) значений для передела 11690 в подчиненной таблице
+     */
     @Query("SELECT th FROM TestHead th WHERE th.conversion.idConversion=11690")
     public List<TestHead> findAllByConversionIdConversion();
 }
